requirements.txt
pip>=24.0.0
requests>=2.31.0
python-dotenv>=1.0.1
google-genai>=0.2.0
jinja2>=3.1.3
pydantic>=2.7.0
